[refactor] extract validation functions to ValidatorUtils class
