+++ SSD_TestShell/SSD_TestShell.vcxproj
@@ -134,16 +134,18 @@
     <ClCompile Include="full_read_command.cpp" />
     <ClCompile Include="full_write_command.cpp" />
     <ClCompile Include="help_command.cpp" />
-    <ClCompile Include="parser.cpp" />
     <ClCompile Include="main.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
+    <ClCompile Include="parser\parser.cpp" />
+    <ClCompile Include="parser\validator.cpp" />
+    <ClCompile Include="parser\validator_utils.cpp" />
     <ClCompile Include="read_command.cpp" />
     <ClCompile Include="test\test_command_processer.cpp">
       <AdditionalOptions Condition="'$(Configuration)|$(Platform)'=='Debug|x64'"> %(AdditionalOptions)</AdditionalOptions>
     </ClCompile>
     <ClCompile Include="test\parser_test.cpp">
-      <AdditionalIncludeDirectories Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">$(MSBuildThisFileDirectory)..\..\lib\native\include\;%(AdditionalIncludeDirectories);$(SolutionDir)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">$(MSBuildThisFileDirectory)..\..\lib\native\include\;%(AdditionalIncludeDirectories);$(SolutionDir)\parser</AdditionalIncludeDirectories>
     </ClCompile>
     <ClCompile Include="write_command.cpp" />
     <ClCompile Include="test\mock_command_processer.h" />
@@ -157,7 +159,9 @@
     <ClInclude Include="full_read_command.h" />
     <ClInclude Include="full_write_command.h" />
     <ClInclude Include="help_command.h" />
-    <ClInclude Include="parser.h" />
+    <ClInclude Include="parser\parser.h" />
+    <ClInclude Include="parser\validator.h" />
+    <ClInclude Include="parser\validator_utils.h" />
     <ClInclude Include="read_command.h" />
     <ClInclude Include="write_command.h" />
   </ItemGroup>
