+++ SSD_TestShell/parser/parser.h
@@ -10,7 +10,4 @@ public:
 
 private:
 	std::vector<std::string> split(const std::string& input);
-	void checkTokenCount(std::vector<std::string>& result, int expectedCounts);
-	void validateAddress(const std::string& address);
-	void validateDataValue(const std::string& dataValue);
 };
